Separate the functions which require edge runtime to be exported via /edge subpath.
